constant names equal to original V

+++ ports/Emulator/projects/test/00_c_type_names.v
@@ -1,13 +1,13 @@
 __global (
-	w_ = 5.0
+	w_ = 5
 )
-const y_ = 5.0 
+const y_ = 5 
 
-fn sum(a_ f64, b_ f64) f64 {
+fn sum(a_ int, b_ int) int {
 	c_ := a_ + b_
 	return c_
 }
 
-x_ := 5.0
+x_ := 5
 
 z_ := sum(x_, w_) + y_
